backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
geoalchemy2==0.14.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
httpx==0.25.1
psycopg2-binary==2.9.9
aiosmtplib==3.0.1
cryptography==41.0.7
